requirements.txt
kaggle>=1.6.6
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
python-dotenv==1.0.1

